Search for mathematician in Mathematics Genealogy Project

Perform an online search using information about an individual mathematician to find their ID in the Mathematics Genealogy Project.

Usage

search_id(
  family = NULL,
  given = NULL,
  middle = NULL,
  university = NULL,
  year = NULL,
  thesis_keyword = NULL,
  country = NULL,
  discipline = NULL
)

Arguments

family

a character(1) string with the family names.

given

a character(1) string with the given names.

middle

a character(1) string with the collapsed middle name(s).

university

a character(1) string with the University at which PhD studied.

year

a character(1) string or integer(1) with the year of completion.

thesis_keyword

a character(1) string with keyword(s) in the PhD thesis title.

country

a character(1) string with the country of study.

discipline

an integer(1) with the mathematical sub-discipline code.

Value

Data frame containing all matches against the provided search terms, with columns:

id

Mathematician ID (as required by get_genealogy());

name

The full name (surname first) of the mathematician;

university

The institution at which PhD was obtained;

year

The year PhD was completed.

Details

Any one or more of the listed arguments can be provided. This will trigger an online search against the live Mathematics Genealogy Project database, so please be considerate and do not spam queries. All the information returned by a standard search on the website is gathered into a data frame and returned, enabling programmatic access to the data.

If you cannot find the individual you are looking for, it could be that they are not in the Mathematics Genealogy Project database. New data can be submitted by following the instructions in the "How to submit updates" section at https://mathgenealogy.org/submit.php.

References

Jackson, A. (2007). “A Labor of Love: The Mathematics Genealogy Project”, Notices of the AMS, 54(8), 1002-1003. https://www.ams.org/notices/200708/tx070801002p.pdf

Mulcahy, C. (2017). “The Mathematics Genealogy Project Comes of Age at Twenty-one”, Notices of the AMS, 64(5), 466-470. https://www.ams.org/journals/notices/201705/rnoti-p466.pdf

Examples

# Don't test these as rely on internet resources that can be temporarily
# unavailable.
# \donttest{
# Search for the package author
search_id("Aslett", "Louis")
#> Error in (function (cnd) {    {        .__handler_frame__. <- TRUE        .__setup_frame__. <- frame        if (inherits(cnd, "message")) {            except <- c("warning", "error")        }        else if (inherits(cnd, "warning")) {            except <- "error"        }        else {            except <- ""        }    }    while (!is_null(cnd)) {        if (inherits(cnd, "error")) {            out <- handlers[[1L]](cnd)            if (!inherits(out, "rlang_zap"))                 throw(out)        }        inherit <- .subset2(.subset2(cnd, "rlang"), "inherit")        if (is_false(inherit)) {            return()        }        cnd <- .subset2(cnd, "parent")    }})(structure(list(message = "Failed to perform HTTP request.",     trace = structure(list(call = list(pkgdown::build_site_github_pages(new_process = FALSE,         install = FALSE), build_site(pkg, preview = FALSE, install = install,         new_process = new_process, ...), build_site_local(pkg = pkg,         examples = examples, run_dont_run = run_dont_run, seed = seed,         lazy = lazy, override = override, preview = preview,         devel = devel), build_reference(pkg, lazy = lazy, examples = examples,         run_dont_run = run_dont_run, seed = seed, override = override,         preview = FALSE, devel = devel), unwrap_purrr_error(purrr::map(topics,         build_reference_topic, pkg = pkg, lazy = lazy, examples_env = examples_env,         run_dont_run = run_dont_run)), withCallingHandlers(code,         purrr_error_indexed = function(err) {            cnd_signal(err$parent)        }), purrr::map(topics, build_reference_topic, pkg = pkg,         lazy = lazy, examples_env = examples_env, run_dont_run = run_dont_run),         map_("list", .x, .f, ..., .progress = .progress), with_indexed_errors(i = i,             names = names, error_call = .purrr_error_call, call_with_cleanup(map_impl,                 environment(), .type, .progress, n, names, i)),         withCallingHandlers(expr, error = function(cnd) {            if (i == 0L) {            }            else {                message <- c(i = "In index: {i}.")                if (!is.null(names) && !is.na(names[[i]]) &&                   names[[i]] != "") {                  name <- names[[i]]                  message <- c(message, i = "With name: {name}.")                }                else {                  name <- NULL                }                cli::cli_abort(message, location = i, name = name,                   parent = cnd, call = error_call, class = "purrr_error_indexed")            }        }), call_with_cleanup(map_impl, environment(), .type,             .progress, n, names, i), .f(.x[[i]], ...), withCallingHandlers(data_reference_topic(topic,             pkg, examples_env = examples_env, run_dont_run = run_dont_run),             error = function(err) {                cli::cli_abort("Failed to parse Rd in {.file {topic$file_in}}",                   parent = err, call = quote(build_reference()))            }), data_reference_topic(topic, pkg, examples_env = examples_env,             run_dont_run = run_dont_run), run_examples(tags$tag_examples[[1]],             env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),             topic = tools::file_path_sans_ext(topic$file_in),             run_dont_run = run_dont_run), highlight_examples(code,             topic, env = env), downlit::evaluate_and_highlight(code,             fig_save = fig_save_topic, env = eval_env, output_handler = handler),         evaluate::evaluate(code, child_env(env), new_device = TRUE,             output_handler = output_handler), withRestarts(with_handlers({            for (expr in tle$exprs) {                ev <- withVisible(eval(expr, envir))                watcher$capture_plot_and_output()                watcher$print_value(ev$value, ev$visible, envir)            }            TRUE        }, handlers), eval_continue = function() TRUE, eval_stop = function() FALSE),         withRestartList(expr, restarts), withOneRestart(withRestartList(expr,             restarts[-nr]), restarts[[nr]]), doWithOneRestart(return(expr),             restart), withRestartList(expr, restarts[-nr]), withOneRestart(expr,             restarts[[1L]]), doWithOneRestart(return(expr), restart),         with_handlers({            for (expr in tle$exprs) {                ev <- withVisible(eval(expr, envir))                watcher$capture_plot_and_output()                watcher$print_value(ev$value, ev$visible, envir)            }            TRUE        }, handlers), eval(call), eval(call), withCallingHandlers(code,             message = `<fn>`, warning = `<fn>`, error = `<fn>`),         withVisible(eval(expr, envir)), eval(expr, envir), eval(expr,             envir), search_id("Aslett", "Louis"), rlang::try_fetch({            resp <- httr2::req_perform(httr2::req_body_form(httr2::req_headers(httr2::request("https://mathgenealogy.org/query-prep.php"),                 `User-Agent` = get_user_agent()), family_name = family,                 given_name = given, other_names = middle, school = university,                 year = year, thesis = thesis_keyword, country = country,                 msc = discipline))        }, error = function(e) {            if (inherits(e, "httr2_failure")) {                cli::cli_abort(c(x = "Please check your internet connection. If that is working, check whether the Maths Genealogy Project is currently down: {.url https://mathgenealogy.org/} -- if both are ok, please file a detailed reproducible bug report {.url https://github.com/louisaslett/maths.genealogy/issues}"),                   parent = e, call = rlang::caller_env())            }            else if (inherits(e, "httr2_http")) {                cli::cli_abort(c(x = "There appears to be a problem at the Maths Genealogy Project. Please try your search directly at {.url https://mathgenealogy.org/} -- if your search works on the site, then please file a detailed reproducible bug report for this package at {.url https://github.com/louisaslett/maths.genealogy/issues}"),                   parent = e, call = rlang::caller_env())            }            else {                cli::cli_abort(c(x = "Unknown error. Please file a detailed reproducible bug report for this package at {.url https://github.com/louisaslett/maths.genealogy/issues}"),                   parent = e, call = rlang::caller_env())            }        }), tryCatch(withCallingHandlers(expr, condition = function(cnd) {            {                .__handler_frame__. <- TRUE                .__setup_frame__. <- frame                if (inherits(cnd, "message")) {                  except <- c("warning", "error")                }                else if (inherits(cnd, "warning")) {                  except <- "error"                }                else {                  except <- ""                }            }            while (!is_null(cnd)) {                if (inherits(cnd, "error")) {                  out <- handlers[[1L]](cnd)                  if (!inherits(out, "rlang_zap"))                     throw(out)                }                inherit <- .subset2(.subset2(cnd, "rlang"), "inherit")                if (is_false(inherit)) {                  return()                }                cnd <- .subset2(cnd, "parent")            }        }), stackOverflowError = handlers[[1L]]), tryCatchList(expr,             classes, parentenv, handlers), tryCatchOne(expr,             names, parentenv, handlers[[1L]]), doTryCatch(return(expr),             name, parentenv, handler), withCallingHandlers(expr,             condition = function(cnd) {                {                  .__handler_frame__. <- TRUE                  .__setup_frame__. <- frame                  if (inherits(cnd, "message")) {                    except <- c("warning", "error")                  }                  else if (inherits(cnd, "warning")) {                    except <- "error"                  }                  else {                    except <- ""                  }                }                while (!is_null(cnd)) {                  if (inherits(cnd, "error")) {                    out <- handlers[[1L]](cnd)                    if (!inherits(out, "rlang_zap"))                       throw(out)                  }                  inherit <- .subset2(.subset2(cnd, "rlang"),                     "inherit")                  if (is_false(inherit)) {                    return()                  }                  cnd <- .subset2(cnd, "parent")                }            }), httr2::req_perform(httr2::req_body_form(httr2::req_headers(httr2::request("https://mathgenealogy.org/query-prep.php"),             `User-Agent` = get_user_agent()), family_name = family,             given_name = given, other_names = middle, school = university,             year = year, thesis = thesis_keyword, country = country,             msc = discipline)), handle_resp(req, resp, error_call = error_call),         cnd_signal(resp)), parent = c(0L, 1L, 2L, 3L, 4L, 5L,     4L, 7L, 8L, 9L, 8L, 8L, 12L, 12L, 14L, 15L, 16L, 17L, 18L,     19L, 20L, 21L, 20L, 23L, 24L, 18L, 26L, 27L, 26L, 18L, 18L,     31L, 32L, 33L, 34L, 35L, 36L, 37L, 34L, 33L, 40L, 41L), visible = c(TRUE,     TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE,     TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE,     TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE,     TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, FALSE,     FALSE), namespace = c("pkgdown", "pkgdown", "pkgdown", "pkgdown",     "pkgdown", "base", "purrr", "purrr", "purrr", "base", "purrr",     "pkgdown", "base", "pkgdown", "pkgdown", "pkgdown", "downlit",     "evaluate", "base", "base", "base", "base", "base", "base",     "base", "evaluate", "base", "base", "base", "base", "base",     "base", "maths.genealogy", "rlang", "base", "base", "base",     "base", "base", "httr2", "httr2", "rlang"), scope = c("::",     "::", ":::", "::", ":::", "::", "::", ":::", ":::", "::",     ":::", "local", "::", ":::", ":::", ":::", "::", "::", "::",     "local", "local", "local", "local", "local", "local", ":::",     "::", "::", "::", "::", "::", "::", "::", "::", "::", "local",     "local", "local", "::", "::", ":::", "::")), row.names = c(NA,     -42L), version = 2L, class = c("rlang_trace", "rlib_trace",     "tbl", "data.frame")), parent = structure(list(message = "SSL peer certificate or SSH remote key was not OK [mathgenealogy.org]:\nSSL certificate problem: unable to get local issuer certificate",         call = curl::curl_fetch_memory(url, handle)), class = c("curl_error_peer_failed_verification",     "curl_error", "error", "condition")), call = httr2::req_perform(httr2::req_body_form(httr2::req_headers(httr2::request("https://mathgenealogy.org/query-prep.php"),         `User-Agent` = get_user_agent()), family_name = family,         given_name = given, other_names = middle, school = university,         year = year, thesis = thesis_keyword, country = country,         msc = discipline)), request = structure(list(url = "https://mathgenealogy.org/query-prep.php",         method = NULL, headers = structure(list(`User-Agent` = "R/4.5.1 maths.genealogy/0.1.3"), redact = character(0), class = "httr2_headers"),         body = list(data = list(family_name = structure("Aslett", class = "AsIs"),             given_name = structure("Louis", class = "AsIs")),             type = "form", content_type = "application/x-www-form-urlencoded",             params = list()), fields = list(), options = list(),         policies = list(), state = <environment>), class = "httr2_request")), class = c("httr2_failure", "httr2_error", "rlang_error", "error", "condition"))): ✖ Please check your internet connection. If that is working, check
#>   whether the Maths Genealogy Project is currently down:
#>   <https://mathgenealogy.org/> -- if both are ok, please file a detailed
#>   reproducible bug report
#>   <https://github.com/louisaslett/maths.genealogy/issues>
#> Caused by error in `httr2::req_perform()`:
#> ! Failed to perform HTTP request.
#> Caused by error in `curl::curl_fetch_memory()`:
#> ! SSL peer certificate or SSH remote key was not OK [mathgenealogy.org]:
#> SSL certificate problem: unable to get local issuer certificate

# You may find it easier to directly use the https://mathgenealogy.org/
# website, and extract the "id" from the URL on the page for the mathematician
# of interest.
# }