Mathematical discipline IDs

Map mathematical disciplines to IDs for use in searching for mathematicians.

Usage

disciplines(search = NULL)

Arguments

search

a character(1) string which will search within disciplines. This can be a regular expression search term if desired.

Value

Data frame, with columns:

id

the discipline ID, as required by search_id() when searching for a mathematician within a specific mathematical discipline;

discipline

the name of the discipline classification, per the Mathematics Genealogy Project.

Examples

# Lookup the ID of any discipline involving the partial word "stat"
disciplines("stat")
#>    id                                 discipline
#> 45 62                                 Statistics
#> 54 82 Statistical mechanics, structure of matter

# Use a regular expression to only exactly match the whole word Statistics and nothing else
disciplines("^statistics$")
#>    id discipline
#> 45 62 Statistics

# Use the above to search only for statisticians with the first name Louis
search_id(given = "Louis", discipline = disciplines("^statistics$")$id)
#>        id                     name                              university year
#> 1  171971            Aslett, Louis                 Trinity College, Dublin 2012
#> 2  297113             Berry, Louis Université Paul Sabatier - Toulouse III 2022
#> 3   92213          Buot, Max Louis                  University of Virginia 2003
#> 4  316613 Cammarata, Louis Vincent                      Harvard University 2024
#> 5  289122    Coatrieux, Jean-Louis                  Université de Rennes 1 1973
#> 6   14711              Cote, Louis                     Columbia University 1954
#> 7  168108             Doray, Louis                  University of Waterloo 1992
#> 8  291284             Ellam, Louis                 Imperial College London 2019
#> 9   83157            Gordon, Louis                     Stanford University 1972
#> 10 207143       Hale, Sarah Louise         North Carolina State University 2016
#> 11 229496           Harper, Louise                    Lancaster University 1997
#> 12 281169            Jensen, Louis                       Aarhus University 2021
#> 13  89308            Jensen, Louis                   Iowa State University 1971
#> 14 192420            Kelly, Louise               University of Strathclyde 1996
#> 15  86857           Mariano, Louis              Carnegie Mellon University 2002
#> 16 225334            Mittel, Louis                     Columbia University 2017
#> 17 144847        Radaoroson, Louis         Université du Québec à Montréal 1996
#> 18 298559            Raynal, Louis               Université de Montpellier 2019
#> 19  60083        Soler, Jean-Louis    Université Joseph Fourier Grenoble I 1970
#> 20 165681    Thurstone, Louis Leon               The University of Chicago 1917
#> 21  34388   Yengoyan, Marie-Louise      University of California, Berkeley 1970